requirements.txt
Flask>=2.2
pytesseract>=0.3.10
yake>=0.4.8
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.4.0
